LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/form-submit-cross-frame-expected.txt


FAIL No replace before load, triggered by cross-iframe formElement.submit() assert_equals: frame's after-submit history.length expected 101 but got 100
